chore(refactor): response in a separate module
